+++ database/migrations/2019_08_15_112421_create_activities_table.php
@@ -15,9 +15,9 @@ class CreateActivitiesTable extends Migration
     {
         Schema::create('activities', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->string('type', 50);
-            $table->unsignedInteger('subject_id')->index();
+            $table->unsignedBigInteger('subject_id')->index();
             $table->string('subject_type', 50);
             $table->timestamps();
         });
